Fix broken links after move
