+++ zipkin-finatra/src/main/scala/com/twitter/zipkin/web/App.scala
@@ -95,7 +95,7 @@ class App(config: ZipkinWebConfig, client: gen.ZipkinQuery.FinagledClient) exten
         case _ => {
           client.getTraceSummariesByIds(ids, adjusters).map {
             _.map { summary =>
-              JsonAdapter(ThriftAdapter(summary))
+              JsonQueryAdapter(ThriftQueryAdapter(summary))
             }
           }
         }
